+++ mafiaTelegramBot/src/main/java/co/mafiagame/bot/handler/AskCommandHandler.java
@@ -28,8 +28,8 @@ public class AskCommandHandler extends TelegramCallbackHandler {
 
     @Override
     public void execute(TCallBackQuery callBackQuery) {
-        Integer callerId = callBackQuery.getFrom().getId();
-        Integer roomId = gameContainer.roomOfUser(callerId);
+        Long callerId = callBackQuery.getFrom().getId();
+        Long roomId = gameContainer.roomOfUser(callerId);
         if (Objects.isNull(roomId))
             throw new GameNotStartedYetException();
         Room room = gameContainer.room(roomId);
@@ -38,7 +38,7 @@ public class AskCommandHandler extends TelegramCallbackHandler {
         client.editMessageText(new EditMessageTextRequest()
             .setText(MessageHolder.get(result ? "user.role.is.mafia" : "user.role.is.not.mafia",
                 room.getLang(),
-                room.findPlayer(Integer.valueOf(target))
+                room.findPlayer(Long.valueOf(target))
                     .orElseThrow(IllegalStateException::new).fullName()))
             .setMessageId(callBackQuery.getMessage().getId())
             .setChatId(callBackQuery.getMessage().getChat().getId())
