+++ mafiaTelegramBot/src/main/java/co/mafiagame/bot/handler/StartFinalElectionCommandHandler.java
@@ -41,7 +41,7 @@ public class StartFinalElectionCommandHandler extends TelegramCommandHandler {
         List<Player> players = room.getGame().alivePlayer();
         List<String> options = players.stream()
                 .map(Player::getUserId)
-                .map(Integer::valueOf)
+                .map(Long::valueOf)
                 .map(accountCache::get).map(a ->
                         MessageHolder.get("vote.to", room.getLang()) + " " + a.fullName())
                 .collect(Collectors.toList());
