+++ mafiaTelegramBot/src/main/java/co/mafiagame/bot/telegram/SendMessage.java
@@ -9,18 +9,18 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class SendMessage {
     @JsonProperty("chat_id")
-    private Integer chatId;
+    private Long chatId;
     private String text;
     @JsonProperty("parse_mode")
     private String parseMode;
     @JsonProperty("reply_to_message_id")
-    private Integer replyToMessageId;
+    private Long replyToMessageId;
 
-    public Integer getChatId() {
+    public Long getChatId() {
         return chatId;
     }
 
-    public SendMessage setChatId(Integer chatId) {
+    public SendMessage setChatId(Long chatId) {
         this.chatId = chatId;
         return this;
     }
@@ -43,11 +43,11 @@ public class SendMessage {
         return this;
     }
 
-    public Integer getReplyToMessageId() {
+    public Long getReplyToMessageId() {
         return replyToMessageId;
     }
 
-    public SendMessage setReplyToMessageId(Integer replyToMessageId) {
+    public SendMessage setReplyToMessageId(Long replyToMessageId) {
         this.replyToMessageId = replyToMessageId;
         return this;
     }
